--- a/src/display/displayTemplateMinimal.h
+++ b/src/display/displayTemplateMinimal.h
@@ -113,5 +113,5 @@ void printScreen() {
     // Show heater output in %
     displayProgressbar(pidOutput / 10, 15, 60, 100);
 
-    u8g2.sendBuffer();
+    displayBufferReady = true;
 }