--- a/src/display/displayTemplateScale.h
+++ b/src/display/displayTemplateScale.h
@@ -119,5 +119,5 @@ void printScreen() {
     // Show heater output in %
     displayProgressbar(pidOutput / 10, 30, 60, 98);
 
-    u8g2.sendBuffer();
+    displayBufferReady = true;
 }