--- a/src/display/displayTemplateTempOnly.h
+++ b/src/display/displayTemplateTempOnly.h
@@ -73,5 +73,5 @@ void printScreen() {
 
     displayStatusbar();
 
-    u8g2.sendBuffer();
+    displayBufferReady = true;
 }