--- a/src/mqtt.h
+++ b/src/mqtt.h
@@ -10,10 +10,16 @@
 #include "userConfig.h"
 #include <Arduino.h>
 #include <PubSubClient.h>
+#include <map>
 #include <os.h>
+#include <string>
+
+std::map<std::string, std::string> mqttLastSent;
 
 unsigned long previousMillisMQTT;
 const unsigned long intervalMQTT = 5000;
+const unsigned long intervalMQTTbrew = 500;
+const unsigned long intervalMQTTstandby = 10000;
 
 WiFiClient net;
 PubSubClient mqtt(net);
@@ -29,6 +35,8 @@ char topic_set[256];
 
 unsigned long lastMQTTConnectionAttempt = millis();
 unsigned int MQTTReCnctCount = 0;
+unsigned long previousMqttConnection = millis();
+unsigned long mqttReconnectInterval = 600000; // 10 minutes
 
 extern std::map<const char*, std::function<editable_t*()>, cmp_str> mqttVars;
 extern std::map<const char*, std::function<double()>, cmp_str> mqttSensors;
@@ -56,12 +64,18 @@ void checkMQTT() {
             if (mqtt.connect(hostname, mqtt_username, mqtt_password, topic_will, 0, true, "offline")) {
                 mqtt.subscribe(topic_set);
                 LOGF(DEBUG, "Subscribed to MQTT Topic: %s", topic_set);
+                MQTTReCnctCount = 0; // reset MQTT reconnect count to zero after a successful connection
             } // Try to reconnect to the server; connect() is a blocking
               // function, watch the timeout!
             else {
                 LOGF(DEBUG, "Failed to connect to MQTT due to reason: %i", mqtt.state());
             }
         }
+    }
+    // reset MQTT reconnect count to zero after mqttReconnectInterval so it can try to connect again
+    else if (millis() - previousMqttConnection >= mqttReconnectInterval) {
+        MQTTReCnctCount = 0;
+        previousMqttConnection = millis();
     }
 }
 
@@ -200,68 +214,76 @@ void mqtt_callback(char* topic, byte* data, unsigned int length) {
  */
 int writeSysParamsToMQTT(bool continueOnError = true) {
     unsigned long currentMillisMQTT = millis();
+    unsigned long interval = (machineState == kBrew) ? intervalMQTTbrew : (machineState == kStandby) ? intervalMQTTstandby : intervalMQTT;
 
-    if ((currentMillisMQTT - previousMillisMQTT >= intervalMQTT) && FEATURE_MQTT == 1) {
+    if ((currentMillisMQTT - previousMillisMQTT >= interval) && FEATURE_MQTT == 1) {
         previousMillisMQTT = currentMillisMQTT;
+        mqttUpdateRunning = true;
 
         if (mqtt.connected()) {
             mqtt_publish("status", (char*)"online");
 
             int errorState = 0;
+            char buffer[32]; // shared buffer for snprintf
 
             for (const auto& pair : mqttVars) {
                 editable_t* e = pair.second();
 
                 switch (e->type) {
                     case kFloat:
-                        if (!mqtt_publish(pair.first, number2string(*(float*)e->ptr), true)) {
-                            errorState = mqtt.state();
-                        }
+                        snprintf(buffer, sizeof(buffer), "%.2f", *(float*)e->ptr);
                         break;
                     case kDouble:
-                        if (!mqtt_publish(pair.first, number2string(*(double*)e->ptr), true)) {
-                            errorState = mqtt.state();
-                        }
-                        break;
                     case kDoubletime:
-                        if (!mqtt_publish(pair.first, number2string(*(double*)e->ptr), true)) {
-                            errorState = mqtt.state();
-                        }
+                        snprintf(buffer, sizeof(buffer), "%.2f", *(double*)e->ptr);
                         break;
-
                     case kInteger:
-                        if (!mqtt_publish(pair.first, number2string(*(int*)e->ptr), true)) {
-                            errorState = mqtt.state();
-                        }
+                        snprintf(buffer, sizeof(buffer), "%d", *(int*)e->ptr);
                         break;
-
                     case kUInt8:
-                        if (!mqtt_publish(pair.first, number2string(*(uint8_t*)e->ptr), true)) {
-                            errorState = mqtt.state();
-                        }
+                        snprintf(buffer, sizeof(buffer), "%u", *(uint8_t*)e->ptr);
                         break;
-
                     case kCString:
-                        if (!mqtt_publish(pair.first, number2string(*(char*)e->ptr), true)) {
-                            errorState = mqtt.state();
-                        }
+                        snprintf(buffer, sizeof(buffer), "%s", (char*)e->ptr);
                         break;
+                    default:
+                        continue; // Skip unknown types
                 }
 
-                if (errorState != 0 && !continueOnError) {
-                    // An error occurred and continueOnError is false, return the error state
-                    return errorState;
+                std::string topic = pair.first;
+                std::string value = std::string(buffer);
+
+                if (mqttLastSent[topic] != value) {
+                    if (mqtt_publish(topic.c_str(), buffer, true)) {
+                        mqttLastSent[topic] = value; // Update only if sent successfully
+                    }
+                    else {
+                        errorState = mqtt.state();
+                        if (errorState != 0 && !continueOnError) {
+                            // An error occurred and continueOnError is false, return the error state
+                            return errorState;
+                        }
+                    }
                 }
             }
 
+            // === mqttSensors loop ===
             for (const auto& pair : mqttSensors) {
-                if (!mqtt_publish(pair.first, number2string(pair.second()))) {
-                    errorState = mqtt.state();
-                }
+                snprintf(buffer, sizeof(buffer), "%.2f", pair.second());
+                std::string topic = pair.first;
+                std::string value = std::string(buffer);
 
-                if (errorState != 0 && !continueOnError) {
-                    // An error occurred and continueOnError is false, return the error state
-                    return errorState;
+                if (mqttLastSent[topic] != value) {
+                    if (mqtt_publish(topic.c_str(), buffer)) {
+                        mqttLastSent[topic] = value;
+                    }
+                    else {
+                        errorState = mqtt.state();
+                        if (errorState != 0 && !continueOnError) {
+                            // An error occurred and continueOnError is false, return the error state
+                            return errorState;
+                        }
+                    }
                 }
             }
         }
@@ -480,7 +502,7 @@ DiscoveryObject GenerateNumberDevice(String name, String displayName, int min_va
  * @return 0 if successful, MQTT connection error code if failed to send messages
  */
 int sendHASSIODiscoveryMsg() {
-    // Sensor, number and switch objects which will always be published
+    hassioUpdateRunning = true;
 
     DiscoveryObject machineStateDevice = GenerateSensorDevice("machineState", "Machine State", "", "enum");
     DiscoveryObject actual_temperature = GenerateSensorDevice("temperature", "Boiler Temperature", "°C", "temperature");