--- a/src/standby.h
+++ b/src/standby.h
@@ -18,41 +18,48 @@ unsigned long timeSinceStandbyMillis = 0;
  * @brief Decrements the remaining standby time every second, counting down from the configured duration
  */
 void updateStandbyTimer(void) {
-    unsigned long currentTime = millis();
+    if (standbyModeOn) {
+        unsigned long currentTime = millis();
 
-    if ((standbyModeRemainingTimeMillis != 0) && ((currentTime % 1000) == 0) && (currentTime != lastStandbyTimeMillis)) {
-        unsigned long standbyModeTimeMillis = standbyModeTime * 60 * 1000;
-        long elapsedTime = currentTime - standbyModeStartTimeMillis;
-        lastStandbyTimeMillis = currentTime;
+        if ((standbyModeRemainingTimeMillis != 0) && ((currentTime % 1000) == 0) && (currentTime != lastStandbyTimeMillis)) {
+            unsigned long standbyModeTimeMillis = standbyModeTime * 60 * 1000;
+            long elapsedTime = currentTime - standbyModeStartTimeMillis;
+            lastStandbyTimeMillis = currentTime;
 
-        if (standbyModeTimeMillis > elapsedTime) {
-            standbyModeRemainingTimeMillis = standbyModeTimeMillis - elapsedTime;
+            if (standbyModeTimeMillis > elapsedTime) {
+                standbyModeRemainingTimeMillis = standbyModeTimeMillis - elapsedTime;
 
-            if ((currentTime % 60000) == 0) {
-                LOGF(INFO, "Standby time remaining: %i minutes", standbyModeRemainingTimeMillis / 60000);
+                if ((currentTime % 60000) == 0) {
+                    LOGF(INFO, "Standby time remaining: %i minutes", standbyModeRemainingTimeMillis / 60000);
+                }
+            }
+            else {
+                standbyModeRemainingTimeMillis = 0;
+                LOG(INFO, "Entering standby mode...");
             }
         }
-        else {
-            standbyModeRemainingTimeMillis = 0;
-            LOG(INFO, "Entering standby mode...");
+        else if ((standbyModeRemainingTimeDisplayOffMillis != 0) && ((currentTime % 1000) == 0) && (currentTime != lastStandbyTimeMillis)) {
+            unsigned long standbyModeTimeMillis = (standbyModeTime * 60 * 1000) + (TIME_TO_DISPLAY_OFF * 60 * 1000);
+            long elapsedTime = currentTime - standbyModeStartTimeMillis;
+            lastStandbyTimeMillis = currentTime;
+
+            if (standbyModeTimeMillis > elapsedTime) {
+                standbyModeRemainingTimeDisplayOffMillis = standbyModeTimeMillis - elapsedTime;
+
+                if ((currentTime % 60000) == 0) {
+                    LOGF(INFO, "Standby time until display is turned off: %i minutes", standbyModeRemainingTimeDisplayOffMillis / 60000);
+                }
+            }
+            else {
+                standbyModeRemainingTimeDisplayOffMillis = 0;
+                LOG(INFO, "Turning off display...");
+            }
         }
     }
-    else if ((standbyModeRemainingTimeDisplayOffMillis != 0) && ((currentTime % 1000) == 0) && (currentTime != lastStandbyTimeMillis)) {
-        unsigned long standbyModeTimeMillis = (standbyModeTime * 60 * 1000) + (TIME_TO_DISPLAY_OFF * 60 * 1000);
-        long elapsedTime = currentTime - standbyModeStartTimeMillis;
-        lastStandbyTimeMillis = currentTime;
-
-        if (standbyModeTimeMillis > elapsedTime) {
-            standbyModeRemainingTimeDisplayOffMillis = standbyModeTimeMillis - elapsedTime;
-
-            if ((currentTime % 60000) == 0) {
-                LOGF(INFO, "Standby time until display is turned off: %i minutes", standbyModeRemainingTimeDisplayOffMillis / 60000);
-            }
-        }
-        else {
-            standbyModeRemainingTimeDisplayOffMillis = 0;
-            LOG(INFO, "Turning off display...");
-        }
+    else {
+        standbyModeRemainingTimeMillis = standbyModeTime * 60 * 1000;
+        standbyModeRemainingTimeDisplayOffMillis = TIME_TO_DISPLAY_OFF * 60 * 1000;
+        standbyModeStartTimeMillis = millis();
     }
 }