Changes have been made to minimise overlap of long running tasks and reduce excessive MQTT data. There is more that could be done but this only required minor changes instead of major rewrites. The boolean flags that block processes do increase the latency of these tasks by an extra loop, which is typically 0.5ms. Each long running task is given a unique flag, so blocking of other tasks can be selective on which tasks it will run with within the same loop.

Before
regular spikes over 100ms and even up to 2s when HASSIO discovery runs, as all the other processes were running in the same loop after this. These caused stuttery behaviour in the main screen and during brews and also increased the frequency of missed switch inputs.

After
These changes have resulted in the maximum loop time being typically 35ms (display sendBuffer) with very infrequent HASSIO discovery sends outside of critical times.

MQTT
Each time writeSysParamsToMQTT was called it would send around 70 messages for around 35 parameters, taking an average of 100ms each send. By storing the sent values in an array it reduces to only the changed variables, so send time is more like 15-50ms. Reducing the impact of these calls allows high speed refreshes during brews without significant increase to loop timing. The MQTT send interval has been split into Normal, Brew and Standby.

Display
sendBuffer() takes around 35ms to complete, and drawing the display into the buffer can take a few ms depending on its complexity. Changing to a boolean to flag the buffer is ready, then calling sendBuffer() on the next loop reduces the loop time to around 5ms then 35ms next loop, instead of one loop of 40ms. Standby mode with display off was still sending data to the display that was set to ignore it.

HASSIO discovery
Every 5 minutes HASSIO discovery was sent, taking between 200ms and 2s. If this happens during a brew the screen pauses for this time while the discovery message is sent. The retain flag is set so MQTT stores the discovery messages, so constant sending is not required. Limiting the send to only while not in a brew and only after MQTT disconnection reduces these blocking calls.

Website
The website refresh is only a few ms, blocking its update to only when other tasks have not run in the same loop minimises maximum loop time.

Standby
It would go into standby even if not enabled, so a check of standbyModeOn was added and if false then keep resetting the timers similar to resetStandbyTimer(void) but without the log.

Other changes
WeightSetpoint was not included in HASSIO discovery, and Steam Kp had an incorrect name.