@scottcondie24 2025-05-29
To visually see the stutters I was changing the Hassio discovery time from 300000 (5mins) to 10000 (10s) then starting a full screen brew. The display shows time to 0.1s and it pauses often with the original code.