@scottcondie24 2025-05-30
Debug functions have been added to print long loop times in the console with which activity caused it. This will aid in review but can be removed before approval if required