@scottcondie24 2025-06-02
I added a fix for MQTT error counts not resetting, now it resets the counter to zero after 10mins since last connection, after waking from standby and after a brew is finished